bar charts feature

+++ app1/models.py
@@ -52,7 +52,7 @@ class UserAchievements(models.Model):
 class UserData(models.Model):
     user=models.ForeignKey(User, on_delete=models.CASCADE)
     date=models.DateField()
-    hours=models.IntegerField()
+    hours=models.FloatField()
